python/requirements.txt
aiohttp==3.6.2
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==19.3.0
blis==0.4.1
boto3==1.13.23
botocore==1.16.23
bump2version==1.0.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorclass==2.2.0
colorlog==4.1.0
croniter==0.3.32
cymem==2.0.3
docutils==0.15.2
# en-core-web-sm==2.2.5
fastavro==0.22.3
faust==1.10.4
faust-avro-serializer==1.0.1
faust-bootstrap==1.1.0
faust-s3-backed-serializer==1.0.1
idna==2.9
importlib-metadata==1.6.0
jmespath==0.10.0
mode==4.3.2
more-itertools==8.3.0
multidict==4.7.6
murmurhash==1.0.2
mypy-extensions==0.4.3
natsort==7.0.1
numpy==1.18.5
opentracing==1.3.0
packaging==20.4
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
pytest-asyncio==0.10.0
python-dateutil==2.8.1
python-schema-registry-client==1.2.4
requests==2.22.0
robinhood-aiokafka==1.1.6
s3transfer==0.3.3
six==1.15.0
spacy==2.2.4
srsly==1.0.2
terminaltables==3.1.0
thinc==7.4.0
tqdm==4.46.1
typing-extensions==3.7.4
urllib3==1.25.9
venusian==1.2.0
wasabi==0.6.0
wcwidth==0.2.3
yarl==1.4.2
zipp==3.1.0
